backend/requirements.txt
Django~=4.2.2
djangorestframework~=3.14.0
web3~=6.5.0